webapi/requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
openai==1.79.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
